Move the transactional method `createReport` into another class
* Move helper method for packageMetadata lookup into PackageMetadataRepository

Fixes GH-273
