+++ spring-cloud-skipper-server-core/src/main/java/org/springframework/cloud/skipper/server/repository/PackageMetadataRepositoryImpl.java
@@ -21,6 +21,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.skipper.SkipperException;
 import org.springframework.cloud.skipper.domain.PackageMetadata;
 import org.springframework.cloud.skipper.domain.Repository;
+import org.springframework.util.Assert;
+import org.springframework.util.StringUtils;
 
 /**
  * Implementation for the {@link PackageMetadataRepositoryCustom} methods.
@@ -55,7 +57,8 @@ public class PackageMetadataRepositoryImpl implements PackageMetadataRepositoryC
 				}
 			}
 		}
-		// if no repoId matches, then return the first package that matches (which has the highest api version set).
+		// if no repoId matches, then return the first package that matches (which has the highest
+		// api version set).
 		return packageMetadataList.get(0);
 	}
 
@@ -64,6 +67,24 @@ public class PackageMetadataRepositoryImpl implements PackageMetadataRepositoryC
 		List<PackageMetadata> packageMetadata = this.packageMetadataRepository.findByName(packageName);
 		if (packageMetadata.isEmpty()) {
 			throw new SkipperException(String.format("Can not find a package named '%s'", packageName));
+		}
+		return packageMetadata;
+	}
+
+	@Override
+	public PackageMetadata findByNameAndOptionalVersionRequired(String packageName, String packageVersion) {
+		Assert.isTrue(StringUtils.hasText(packageName), "Package name must not be empty");
+		PackageMetadata packageMetadata;
+		if (StringUtils.hasText(packageVersion)) {
+			packageMetadata = this.packageMetadataRepository.findByNameAndVersionByMaxRepoOrder(packageName,
+					packageVersion);
+		}
+		else {
+			packageMetadata = this.packageMetadataRepository.findFirstByNameOrderByVersionDesc(packageName);
+		}
+		if (packageMetadata == null) {
+			throw new SkipperException(String.format("Can not find package '%s', version '%s'", packageName,
+					packageVersion));
 		}
 		return packageMetadata;
 	}
